requirements.txt
tensorflow==2.10.0
keras==2.10.0
tensorflow-addons==0.22.0
numpy==1.24.2
matplotlib==3.9.4
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
Pillow==11.3.0
opencv-python==4.11.0.86
